Add new ActionButton with loading state to PageActions and the modals
