+++ services/app-web/src/pages/StateSubmission/SubmissionType/SubmissionType.test.tsx
@@ -13,9 +13,10 @@ describe('SubmissionType', () => {
         submissionDescription: '',
         submissionType: '',
     }
+    const updateDraftMock = jest.fn()
 
     it('displays correct form guidance', async () => {
-        renderWithProviders(<SubmissionType />, {
+        renderWithProviders(<SubmissionType updateDraft={updateDraftMock} />, {
             apolloProvider: {
                 mocks: [fetchCurrentUserMock({ statusCode: 200 })],
             },
@@ -25,7 +26,7 @@ describe('SubmissionType', () => {
     })
 
     it('displays submission type form when expected', async () => {
-        renderWithProviders(<SubmissionType />, {
+        renderWithProviders(<SubmissionType updateDraft={updateDraftMock} />, {
             apolloProvider: {
                 mocks: [fetchCurrentUserMock({ statusCode: 200 })],
             },
@@ -39,7 +40,7 @@ describe('SubmissionType', () => {
     })
 
     it('displays new submission form when expected', async () => {
-        renderWithProviders(<SubmissionType />, {
+        renderWithProviders(<SubmissionType updateDraft={updateDraftMock} />, {
             apolloProvider: {
                 mocks: [fetchCurrentUserMock({ statusCode: 200 })],
             },
@@ -57,7 +58,10 @@ describe('SubmissionType', () => {
 
     it('displays with draft submission when expected', async () => {
         renderWithProviders(
-            <SubmissionType draftSubmission={contractOnly()} />,
+            <SubmissionType
+                updateDraft={updateDraftMock}
+                draftSubmission={contractOnly()}
+            />,
             {
                 apolloProvider: {
                     mocks: [fetchCurrentUserMock({ statusCode: 200 })],
@@ -73,7 +77,7 @@ describe('SubmissionType', () => {
     })
 
     it('displays a cancel link', async () => {
-        renderWithProviders(<SubmissionType />, {
+        renderWithProviders(<SubmissionType updateDraft={updateDraftMock} />, {
             apolloProvider: {
                 mocks: [fetchCurrentUserMock({ statusCode: 200 })],
             },
@@ -89,7 +93,7 @@ describe('SubmissionType', () => {
     })
 
     it('displays a continue button', async () => {
-        renderWithProviders(<SubmissionType />, {
+        renderWithProviders(<SubmissionType updateDraft={updateDraftMock} />, {
             apolloProvider: {
                 mocks: [fetchCurrentUserMock({ statusCode: 200 })],
             },
@@ -110,7 +114,7 @@ describe('SubmissionType', () => {
                 initialValues={SubmissionTypeInitialValues}
                 onSubmit={jest.fn()}
             >
-                <SubmissionType />
+                <SubmissionType updateDraft={updateDraftMock} />
             </Formik>,
             {
                 apolloProvider: {
@@ -147,7 +151,7 @@ describe('SubmissionType', () => {
                 initialValues={SubmissionTypeInitialValues}
                 onSubmit={jest.fn()}
             >
-                <SubmissionType />
+                <SubmissionType updateDraft={updateDraftMock} />
             </Formik>,
             {
                 apolloProvider: {
@@ -192,7 +196,7 @@ describe('SubmissionType', () => {
                 initialValues={SubmissionTypeInitialValues}
                 onSubmit={jest.fn()}
             >
-                <SubmissionType />
+                <SubmissionType updateDraft={updateDraftMock} />
             </Formik>,
             {
                 apolloProvider: {
@@ -219,7 +223,7 @@ describe('SubmissionType', () => {
                 initialValues={SubmissionTypeInitialValues}
                 onSubmit={jest.fn()}
             >
-                <SubmissionType />
+                <SubmissionType updateDraft={updateDraftMock} />
             </Formik>,
             {
                 apolloProvider: {
@@ -236,11 +240,14 @@ describe('SubmissionType', () => {
 
     describe('validations', () => {
         it('does not show error validations on initial load', async () => {
-            renderWithProviders(<SubmissionType />, {
-                apolloProvider: {
-                    mocks: [fetchCurrentUserMock({ statusCode: 200 })],
-                },
-            })
+            renderWithProviders(
+                <SubmissionType updateDraft={updateDraftMock} />,
+                {
+                    apolloProvider: {
+                        mocks: [fetchCurrentUserMock({ statusCode: 200 })],
+                    },
+                }
+            )
 
             await waitFor(() => {
                 expect(screen.getByRole('textbox')).not.toHaveClass(
@@ -259,7 +266,10 @@ describe('SubmissionType', () => {
 
         it('shows error messages when there are validation errors and showValidations is true', async () => {
             renderWithProviders(
-                <SubmissionType showValidations={true} />,
+                <SubmissionType
+                    updateDraft={updateDraftMock}
+                    showValidations={true}
+                />,
 
                 {
                     apolloProvider: {
@@ -288,11 +298,17 @@ describe('SubmissionType', () => {
         })
 
         it('do not show error messages when showValidations is false', async () => {
-            renderWithProviders(<SubmissionType showValidations={false} />, {
-                apolloProvider: {
-                    mocks: [fetchCurrentUserMock({ statusCode: 200 })],
-                },
-            })
+            renderWithProviders(
+                <SubmissionType
+                    updateDraft={updateDraftMock}
+                    showValidations={false}
+                />,
+                {
+                    apolloProvider: {
+                        mocks: [fetchCurrentUserMock({ statusCode: 200 })],
+                    },
+                }
+            )
             await waitFor(() => {
                 const textarea = screen.getByRole('textbox', {
                     name: 'Submission description',
@@ -313,11 +329,14 @@ describe('SubmissionType', () => {
 
     describe('Continue / Save Draft button', () => {
         it('if form fields are invalid, shows validation error messages when continue button is clicked', async () => {
-            renderWithProviders(<SubmissionType />, {
-                apolloProvider: {
-                    mocks: [fetchCurrentUserMock({ statusCode: 200 })],
-                },
-            })
+            renderWithProviders(
+                <SubmissionType updateDraft={updateDraftMock} />,
+                {
+                    apolloProvider: {
+                        mocks: [fetchCurrentUserMock({ statusCode: 200 })],
+                    },
+                }
+            )
 
             userEvent.click(
                 screen.getByRole('button', {
